feat(client): teacher preview should be same as student attempting the test [EV-24561] (#14032)

+++ src/client/publicTest/utils.js
@@ -145,7 +145,15 @@ export const showTestInfoModal = ({
   history,
   title,
   notifyCancel,
+  closeTestPreviewModal,
+  preview,
 }) => {
+  let selectedLang = ''
+  const handlChange = (value) => {
+    setSelectedLanguage(value)
+    selectedLang = value
+  }
+
   const timedContent = pauseAllowed ? (
     <p style={{ margin: '10px 0' }}>
       {' '}
@@ -182,9 +190,9 @@ export const showTestInfoModal = ({
           <p style={{ margin: '10px 0' }}>
             <Select
               getPopupContainer={(e) => e.parentElement}
-              defaultValue={languagePreference}
+              defaultValue={languagePreference || ''}
               style={{ width: 200 }}
-              onChange={setSelectedLanguage}
+              onChange={handlChange}
               suffixIcon={<IconSelectCaretDown color={themeColor} />}
             >
               <Option value="" disabled>
@@ -225,20 +233,34 @@ export const showTestInfoModal = ({
     content,
     onOk: () => {
       if (attemptCount < maxAttempts)
-        startAssignment({ testId, assignmentId, testType, classId })
-      Modal.destroyAll()
+        startAssignment({
+          testId,
+          assignmentId,
+          testType,
+          classId,
+          selectedLang,
+        })
+      if (!preview) Modal.destroyAll()
+      if (preview && multiLanguageEnabled) {
+        return !selectedLang
+      }
     },
     onCancel: () => {
       setSelectedLanguage('')
       if (notifyCancel) redirectToDashbord('HOME', history)
       else Modal.destroyAll()
+      if (preview) {
+        closeTestPreviewModal()
+      }
     },
     okText: 'YES, CONTINUE',
     cancelText: 'NO, CANCEL',
     className: 'ant-modal-confirm-custom-styled',
     centered: true,
+    maskClosable: !preview,
     icon: '',
   })
+  return null
 }
 
 // case: check to where to navigate
